rename to wrench, add Makefile
